Generate DID with method pattern
Change method name to method pattern for key-to-did and
key-to-verification-method.

Method pattern is a generalization of method name that allows passing an
additional string to the DID method by appending it to the DID method
name, separated by a colon (':'). This can be used to request generating
a DID under a given prefix or other pattern if supported by the DID
method.